add float32 to optimizer

+++ optimizers.py
@@ -80,22 +80,22 @@ def adam(lr, tparams, grads, inp, cost, hard_attn_up):
     gsup = [(gs, g) for gs, g in zip(gshared, grads)]
 
     f_grad_shared = theano.function(inp, cost, updates=gsup+hard_attn_up)
-    lr0 = 0.0002
-    b1 = 0.1
-    b2 = 0.001
-    e = 1e-8
+    lr0 = numpy.float32(0.0002)
+    b1 = numpy.float32(0.1)
+    b2 = numpy.float32(0.001)
+    e = numpy.float32(1e-8)
     updates = []
     i = theano.shared(numpy.float32(0.))
-    i_t = i + 1.
-    fix1 = 1. - b1**(i_t)
-    fix2 = 1. - b2**(i_t)
+    i_t = i + numpy.float32(1.)
+    fix1 = numpy.float32(1.) - b1**(i_t)
+    fix2 = numpy.float32(1.) - b2**(i_t)
     lr_t = lr0 * (tensor.sqrt(fix2) / fix1)
 
     for p, g in zip(tparams.values(), gshared):
         m = theano.shared(p.get_value() * numpy.float32(0.))
         v = theano.shared(p.get_value() * numpy.float32(0.))
-        m_t = (b1 * g) + ((1. - b1) * m)
-        v_t = (b2 * tensor.sqr(g)) + ((1. - b2) * v)
+        m_t = (b1 * g) + ((numpy.float32(1.) - b1) * m)
+        v_t = (b2 * tensor.sqr(g)) + ((numpy.float32(1.) - b2) * v)
         g_t = m_t / (tensor.sqrt(v_t) + e)
         p_t = p - (lr_t * g_t)
         updates.append((m, m_t))
